Allow skipping all refetching when performing mutation
Fixes https://github.com/vadimdemedes/draqula/issues/18

There are cases when it's not important enough for a mutation
to refetch queries. For example, a like counter component
will most likely keep track of likes in a local state to ensure
instant response on click to increment a counter. In that case,
it's only important to increment like counter on backend and
not spend bandwidth on unnecessarily refreshing other parts of the page.

This change allows `refetchQueries` to be set to `false` to disable
any kind of refetching