Fix race condition when fetching more items after page is focused
`fetchMore()` and `refetch()` were triggered at the same time, which resulted
in new items being merged with out-dated items, if `setData()` was called to
manually update the `data` object.

This commit fixes it by replacing `useReducer()` with `useState()` + callbacks,
which ensures all state updates are queued and state is changed sequentially,
based on the last value.